+++ NEWS.md
@@ -1,6 +1,7 @@
 # CAISER 0.2.3
 * Added new example and use case to `calc_nreps2()`
 * Improved plot and summary functions for `CAISErPowercurve` objects.
+* Added options to `calc_power_curve()` to determine the range of effect sizes to consider.
 
 # CAISEr 0.2.1
 * Minor fixes, particularly in printing function.